Feature: layout파일을 SiteFrame으로 변경 관련 파일 수정
