+++ src/@types/views/index.d.ts
@@ -1,9 +1,6 @@
 declare module "view" {
 
-  import { ReactChild } from "react"
-
   export interface BaseViewProps {
     className?: string
-    children?: ReactChild
   }
 }
