+++ src/components/organisms/Header/header.scss
@@ -19,5 +19,5 @@
 }
 
 .mdc-top-app-bar.mdc-top-app-bar--short-collapsed {
-  width: 8rem;
+  width: 9.2rem;
 }
