+++ src/components/templates/SiteFrame/layout.scss
@@ -1,0 +1,6 @@
+.content-wrapper {
+  margin: 56px auto;
+  max-width: 960px;
+  height: 1500px;
+  padding: 0 1.0875rem 1.45rem;
+}
