+++ src/pages/404.js
@@ -1,14 +1,14 @@
 import React from "react"
 
-import Layout from "../components/templates/layout"
+import Frame from "../components/templates/SiteFrame"
 import SEO from "../components/seo"
 
 const NotFoundPage = () => (
-  <Layout>
+  <Frame>
     <SEO title="404: Not found" />
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </Layout>
+  </Frame>
 )
 
 export default NotFoundPage
