+++ src/pages/index.js
@@ -1,23 +1,23 @@
 import React from "react"
 import { Link } from "gatsby"
 
-import Layout from "../components/templates/layout"
-import Index from "../components/atoms/Image"
+import Frame from "../components/templates/SiteFrame"
+import Image from "../components/atoms/Image"
 import SEO from "../components/seo"
 
 import "./global.scss"
 
 const IndexPage = () => (
-  <Layout>
+  <Frame>
     <SEO title="Home"/>
     <h1>Hi people</h1>
     <p>Welcome to your new Gatsby site.</p>
     <p>Now go build something great.</p>
     <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
-      <Index/>
+      <Image/>
     </div>
     <Link to="/page-2/">Go to page 2</Link>
-  </Layout>
+  </Frame>
 )
 
 export default IndexPage
