+++ src/pages/page-2.tsx
@@ -1,17 +1,16 @@
-// Gatsby supports TypeScript natively!
 import React from "react"
 import { PageProps, Link } from "gatsby"
 
-import Layout from "../components/templates/layout"
+import Frame from "../components/templates/SiteFrame"
 import SEO from "../components/seo"
 
 const SecondPage = (props: PageProps) => (
-  <Layout>
+  <Frame>
     <SEO title="Page two" />
     <h1>Hi from the second page</h1>
     <p>Welcome to page 2 ({props.path})</p>
     <Link to="/">Go back to the homepage</Link>
-  </Layout>
+  </Frame>
 )
 
 export default SecondPage
